modularized extract logic in python
pushing to master directly. added cleansing layer and fact table to dwh

added workout loader function

modularized extract logic in python
